Fix "file_exists(): Passing null to parameter of type string is deprecated" (#1148)

+++ application/modules/events/mappers/Events.php
@@ -382,7 +382,7 @@ class Events extends \Ilch\Mapper
             ->execute()
             ->fetchAssoc();
 
-        if (file_exists($imageRow['image'])) {
+        if (isset($imageRow['image']) && file_exists($imageRow['image'])) {
             unlink($imageRow['image']);
         }
